\name{IC}
\docType{data}	
\alias{IC}
\alias{ICshumanBPall}
\alias{ICshumanCCall}
\alias{ICshumanMFall}
\title{Information content of GO terms}
\description{  
  \describe{
  \item{"ICshumanBPall"}{Information content of GO terms in "biological process" using all evidence codes for human}
  \item{"ICshumanCCall"}{Information content of GO terms in "cellular component" using all evidence codes for human}  
  \item{"ICshumanMFall"}{Information content of GO terms in "molecular function" using all evidence codes for human}
  }  
}	
\format{A vector of double values}	
\note{
The currently used IC values can be accessed within the GOSimEnv environment.
}	
\keyword{datasets}
