\name{IC}
\docType{data}	
\alias{IC}
\alias{ICsBP}	
\alias{ICsBPIMP_IGI_IDA_IEP_IPI}
\alias{ICsCCall}
\alias{ICsCCIMP_IGI_IDA_IEP_IPI}
\alias{ICsMFall}
\alias{ICsMFIMP_IGI_IDA_IEP_IPI}	
\title{Information content of GO terms}
\description{
  \begin{ldescription}
  \item{"ICsBP"}{Information content of GO terms in "biological process" using all evidence codes}
  \item{"ICsBPIMP\_IGI\_IDA\_IEP\_IPI"}{Information content of GO terms in "biological process" using evidence codes "IMP, IGI, IDA, IEP, IPI"}
  \item{"ICsCCall"}{Information content of GO terms in "cellular component" using all evidence codes}
  \item{"ICsCCIMP\_IGI\_IDA\_IEP\_IPI"}{Information content of GO terms in "cellular component" using evidence codes "IMP, IGI, IDA, IEP, IPI"}
  \item{"ICsMFall"}{Information content of GO terms in "molecular function" using all evidence codes}
  \item{"ICsMFIMP\_IGI\_IDA\_IEP\_IPI"}{Information content of GO terms in "molecular function" using evidence "IMP, IGI, IDA, IEP, IPI"}
  \end{ldescription}
  
}	
\format{A vector of double values}		
\keyword{datasets}
