\name{IC}
\docType{data}	
\alias{IC}
\alias{ICshumanBPall}
\alias{ICshumanCCall}
\alias{ICshumanMFall}
\alias{ICsmouseBPall}
\alias{ICsmouseCCall}
\alias{ICsmouseMFall}
\alias{ICsratBPall}
\alias{ICsratCCall}
\alias{ICsratMFall}
\alias{ICsyeastBPall}
\alias{ICsyeastCCall}
\alias{ICsyeastMFall}
\alias{ICsflyBPall}
\alias{ICsflyCCall}
\alias{ICsflyMFall}
\alias{ICsmalariaBPall}
\alias{ICsmalariaCCall}
\alias{ICsmalariaMFall}

\title{Information content of GO terms}
\description{  
  \item{"ICshumanBPall"}{Information content of GO terms in "biological process" using all evidence codes for human}
  \item{"ICshumanCCall"}{Information content of GO terms in "cellular component" using all evidence codes for human}  
  \item{"ICshumanMFall"}{Information content of GO terms in "molecular function" using all evidence codes for human}
  
  \item{"ICsyeastBPall"}{Information content of GO terms in "biological process" using all evidence codes for yeast}
  \item{"ICsyeastCCall"}{Information content of GO terms in "cellular component" using all evidence codes for yeast}  
  \item{"ICsyeastMFall"}{Information content of GO terms in "molecular function" using all evidence codes for yeast}  

  \item{"ICsflyBPall"}{Information content of GO terms in "biological process" using all evidence codes for fly}
  \item{"ICsflyCCall"}{Information content of GO terms in "cellular component" using all evidence codes for fly}  
  \item{"ICsflyMFall"}{Information content of GO terms in "molecular function" using all evidence codes for fly}  

  \item{"ICsmalariaBPall"}{Information content of GO terms in "biological process" using all evidence codes for malaria}
  \item{"ICsmalariaCCall"}{Information content of GO terms in "cellular component" using all evidence codes for malaria}  
  \item{"ICsmalariaMFall"}{Information content of GO terms in "molecular function" using all evidence codes for malaria}  

  \item{"ICsratBPall"}{Information content of GO terms in "biological process" using all evidence codes for rat}
  \item{"ICsratCCall"}{Information content of GO terms in "cellular component" using all evidence codes for rat}  
  \item{"ICsratMFall"}{Information content of GO terms in "molecular function" using all evidence codes for rat}  

  \item{"ICsmouseBPall"}{Information content of GO terms in "biological process" using all evidence codes for mouse}
  \item{"ICsmouseCCall"}{Information content of GO terms in "cellular component" using all evidence codes for mouse}  
  \item{"ICsmouseMFall"}{Information content of GO terms in "molecular function" using all evidence codes for mouse}  
}	
\format{A vector of double values}	
\note{
The currently used IC values can be accessed within the GOSimEnv environment.
}	
\keyword{datasets}
