\name{internal}
\alias{internal}
\alias{calcTermSim}
\alias{getDensityFactor}
\alias{getDepthFactor}
\alias{getDisjAnc}
\alias{getDisjCommAncSim}
\alias{getEnrichedSim}
\alias{getGSim}
\alias{precomputeTermSims}
\alias{initialize}
\alias{pca}
\alias{norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{  internal functions  }
\description{
  internal functions: do not call these functions directly. 
}
\usage{
	various
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	various
}
\details{
}
\value{
 	various
}
\author{ Holger Froehlich }
\keyword{file}

